computing damage state betas
