+++ sc/app/aebm/damage.py
@@ -1,1358 +1,5 @@
-# interstory drift ratio
-interstory_drift = {
-  'special_high_baseline': {
-    'C1': {
-      'slight': .0063,
-      'moderate': .0125,
-      'extensive': .038,
-      'complete': .1
-    },
-    'C2': {
-      'slight': .005,
-      'moderate': .0125,
-      'extensive': .038,
-      'complete': .1
-    },
-    'C3': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .005,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .0875
-    },
-    'PC2': {
-      'slight': .005,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .0875
-    },
-    'RM1': {
-      'slight': .005,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .0875
-    },
-    'RM2': {
-      'slight': .005,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .088
-    },
-    'S1': {
-      'slight': .0075,
-      'moderate': .015,
-      'extensive': .0375,
-      'complete': .1 
-    },
-    'S2': {
-      'slight': .0063,
-      'moderate': .013,
-      'extensive': .0375,
-      'complete': .1 
-    },
-    'S3': {
-      'slight': .005,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .0875 
-    },
-    'S4': {
-      'slight': .005,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .0875
-    },
-    'S5': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'URM': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'W1': {
-      'slight': .005,
-      'moderate': .015,
-      'extensive': .05,
-      'complete': .125
-    },
-    'W1A': {
-      'slight': .005,
-      'moderate': .015,
-      'extensive': .05,
-      'complete': .125
-    },
-    'W2': {
-      'slight': .005,
-      'moderate': .015,
-      'extensive': .05,
-      'complete': .125
-    }
-  },
-  'high_baseline': {
-    'C1': {
-      'slight': .005,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .8
-    },
-    'C2': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .8
-    },
-    'C3': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .024,
-      'complete': .07
-    },
-    'PC2': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .024,
-      'complete': .07
-    },
-    'RM1': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .024,
-      'complete': .07
-    },
-    'RM2': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .024,
-      'complete': .07
-    },
-    'S1': {
-      'slight': .006,
-      'moderate': .012,
-      'extensive': .03,
-      'complete': .08
-    },
-    'S2': {
-      'slight': .005,
-      'moderate': .01,
-      'extensive': .03,
-      'complete': .08
-    },
-    'S3': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .024,
-      'complete': .07 
-    },
-    'S4': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .024,
-      'complete': .07
-    },
-    'S5': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'URM': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'W1': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'W1A': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'W2': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    }
-  },
-  'moderate_baseline': {
-    'C1': {
-      'slight': .005,
-      'moderate': .009,
-      'extensive': .023,
-      'complete': .06
-    },
-    'C2': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .023,
-      'complete': .06
-    },
-    'C3': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'PC2': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'RM1': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'RM2': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'S1': {
-      'slight': .006,
-      'moderate': .01,
-      'extensive': .024,
-      'complete': .06
-    },
-    'S2': {
-      'slight': .005,
-      'moderate': .009,
-      'extensive': .023,
-      'complete': .06
-    },
-    'S3': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'S4': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'S5': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'URM': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'W1': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .031,
-      'complete': .075
-    },
-    'W1A': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .031,
-      'complete': .075
-    },
-    'W2': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .031,
-      'complete': .075
-    }
-  },
-  'high_poor': {
-    'C1': {
-      'slight': .005,
-      'moderate': .009,
-      'extensive': .023,
-      'complete': .06
-    },
-    'C2': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .023,
-      'complete': .06
-    },
-    'C3': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'PC2': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'RM1': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'RM2': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'S1': {
-      'slight': .006,
-      'moderate': .01,
-      'extensive': .024,
-      'complete': .06
-    },
-    'S2': {
-      'slight': .005,
-      'moderate': .009,
-      'extensive': .023,
-      'complete': .06
-    },
-    'S3': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'S4': {
-      'slight': .004,
-      'moderate': .007,
-      'extensive': .019,
-      'complete': .053
-    },
-    'S5': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'URM': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'W1': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .031,
-      'complete': .075
-    },
-    'W1A': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .031,
-      'complete': .075
-    },
-    'W2': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .031,
-      'complete': .075
-    }
-  },
-  'low_baseline': {
-    'C1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'C2': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'C3': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'PC2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'RM1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'RM2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'S1': {
-      'slight': .006,
-      'moderate': .01,
-      'extensive': .02,
-      'complete': .05
-    },
-    'S2': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'S3': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'S4': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'S5': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'URM': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'W1': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .025,
-      'complete': .06
-    },
-    'W1A': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .025,
-      'complete': .06
-    },
-    'W2': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .025,
-      'complete': .06
-    },
-  },
-  'moderate_poor': {
-    'C1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'C2': {
-      'slight': .004,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'C3': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'PC2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'RM1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'RM2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'S1': {
-      'slight': .006,
-      'moderate': .01,
-      'extensive': .02,
-      'complete': .05
-    },
-    'S2': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'S3': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'S4': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .044
-    },
-    'S5': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'URM': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .015,
-      'complete': .035
-    },
-    'W1': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .025,
-      'complete': .06
-    },
-    'W1A': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .025,
-      'complete': .06
-    },
-    'W2': {
-      'slight': .004,
-      'moderate': .01,
-      'extensive': .025,
-      'complete': .06
-    }
-  },
-  'pre_baseline': {
-    'C1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'C2': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'C3': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'PC2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'RM1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'RM2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .016,
-      'complete': .04
-    },
-    'S2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'S3': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S4': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S5': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'URM': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'W1': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'W1A': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'W2': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    }
-  },
-  'low_poor': {
-    'C1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'C2': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'C3': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'PC2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'RM1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'RM2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .016,
-      'complete': .04
-    },
-    'S2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'S3': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S4': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S5': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'URM': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'W1': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'W1A': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'W2': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    }
-  },
-  'high_very_poor': {
-    'C1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'C2': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'C3': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'PC2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'RM1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'RM2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .016,
-      'complete': .04
-    },
-    'S2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .016,
-      'complete': .04
-    },
-    'S3': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S4': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .013,
-      'complete': .035
-    },
-    'S5': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'URM': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .012,
-      'complete': .028
-    },
-    'W1': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'W1A': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    },
-    'W2': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .02,
-      'complete': .05
-    }
-  },
-  'pre_poor': {
-    'C1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .0115,
-      'complete': .03
-    },
-    'C2': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .0115,
-      'complete': .03
-    },
-    'C3': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .0175
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265
-    },
-    'PC2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265
-    },
-    'RM1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265
-    },
-    'RM2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .027
-    },
-    'S1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .012,
-      'complete':.03 
-    },
-    'S2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .012,
-      'complete':.03 
-    },
-    'S3': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265 
-    },
-    'S4': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265
-    },
-    'S5': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .0175
-    },
-    'URM': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .0175
-    },
-    'W1': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .018,
-      'complete': .045
-    },
-    'W1A': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .018,
-      'complete': .045
-    },
-    'W2': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .018,
-      'complete': .045
-    }
-  },
-  'moderate_very_poor': {
-    'C1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .0115,
-      'complete': .03
-    },
-    'C2': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .0115,
-      'complete': .03
-    },
-    'C3': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .0175
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265
-    },
-    'PC2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265
-    },
-    'RM1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265
-    },
-    'RM2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .027
-    },
-    'S1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .012,
-      'complete':.03 
-    },
-    'S2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .012,
-      'complete':.03 
-    },
-    'S3': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265 
-    },
-    'S4': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .0095,
-      'complete': .0265
-    },
-    'S5': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .0175
-    },
-    'URM': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .0175
-    },
-    'W1': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .018,
-      'complete': .045
-    },
-    'W1A': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .018,
-      'complete': .045
-    },
-    'W2': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .018,
-      'complete': .045
-    }
-  },
-  'low_very_poor': {
-    'C1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .01,
-      'complete': .025
-    },
-    'C2': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .01,
-      'complete': .025
-    },
-    'C3': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .0175
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .022
-    },
-    'PC2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .022
-    },
-    'RM1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .022
-    },
-    'RM2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .022
-    },
-    'S1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .01,
-      'complete': .025
-    },
-    'S2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .01,
-      'complete': .025
-    },
-    'S3': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .022
-    },
-    'S4': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .022
-    },
-    'S5': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .0175
-    },
-    'URM': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .0175
-    },
-    'W1': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .015,
-      'complete': .0375
-    },
-    'W1A': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .015,
-      'complete': .0375
-    },
-    'W2': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .015,
-      'complete': .0375
-    }
-  },
-  'pre_very_poor': {
-    'C1': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .01,
-      'complete': .02
-    },
-    'C2': {
-      'slight': .003,
-      'moderate': .006,
-      'extensive': .01,
-      'complete': .02
-    },
-    'C3': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .014
-    },
-    'MH': {
-      'slight': .004,
-      'moderate': .012,
-      'extensive': .04,
-      'complete': .1
-    },
-    'PC1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .0175
-    },
-    'PC2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .0175
-    },
-    'RM1': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .0175
-    },
-    'RM2': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .0175
-    },
-    'S1': {
-      'slight': .005,
-      'moderate': .008,
-      'extensive': .01,
-      'complete': .02
-    },
-    'S2': {
-      'slight': .004,
-      'moderate': .006,
-      'extensive': .01,
-      'complete': .02
-    },
-    'S3': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .0175
-    },
-    'S4': {
-      'slight': .003,
-      'moderate': .005,
-      'extensive': .008,
-      'complete': .0175
-    },
-    'S5': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .014
-    },
-    'URM': {
-      'slight': .002,
-      'moderate': .005,
-      'extensive': .0075,
-      'complete': .014
-    },
-    'W1': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .015,
-      'complete': .03
-    },
-    'W1A': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .015,
-      'complete': .03
-    },
-    'W2': {
-      'slight': .003,
-      'moderate': .008,
-      'extensive': .015,
-      'complete': .03
-    }
-  }
-}
+from data_tables import interstory_drift, default_beta
+import math
 
 def get_drift(mbt, sdl, perf, state):
     return interstory_drift['{}_{}'.format(sdl, perf)][mbt][state]
@@ -1380,9 +27,138 @@ def get_damage_state_medians(mbt, sdl, performance_rating, height, modal_height,
             drift * height * 12 * (modal_height / modal_response_)
         )
 
-    '''
-    beta
-    
-    =IF(OR($BB5="",$AU5="",$AW5="",$AF5="",$AI5="",$AL5="",$AO5=""),"",MIN(1.1*$BB5,MAX(0.9*$BB5,SQRT((MIN(MAX(LN(IF($AI5<1.2*$BM5,$AI5,1.2*$BM5)/IF($AL5<1.2*$BM5,$AL5,1.2*$BM5))/2,$AV5/2),2*$AV5))^2+(MIN(MAX(LN(IF($AJ5<1.2*$BM5,$AJ5,1.2*$BM5)/IF($AM5<1.2*$BM5,$AM5,1.2*$BM5))/2,$AU5/2),2*$AU5))^2+$AW5^2))))
-    '''
     return states
+
+def get_default_damage_state_beta(quality_rating, performance_rating, year, stories):
+
+    if year > 1975:
+        year_cat = 'post-1975'
+    elif year > 1960:
+        year_cat = '1960-1975'
+    elif year > 1940:
+        year_cat = '1941-1960'
+    else:
+        year_cat = 'pre-1941'
+
+    performance_rating = (
+        performance_rating if performance_rating == 'baseline' 
+        else 'non-baseline'
+    )
+
+    stories = stories if stories <= 15 else 15
+
+    return default_beta[quality_rating][performance_rating][year_cat][stories]
+    
+
+def get_damage_state_beta(default_beta, default_median, lower_bound_demand_disp, lower_bound_demand_acc, upper_bound_demand_disp, upper_bound_demand_acc, demand_uncertainty, quality_rating, performance_rating, year, stories):
+
+    '''
+        MIN(
+            1.1*default_beta,
+            MAX(
+                0.9*default_beta,
+                SQRT(
+                    (MIN(
+                        MAX(
+                            LN(
+                                IF(
+                                    upper_bound_demand_disp<1.2*default_median,
+                                    upper_bound_demand_disp,
+                                    1.2*default_median
+                                )/
+                                IF(lower_bound_demand_disp<1.2*default_median,
+                                lower_bound_demand_disp,
+                                1.2*default_median)
+                            ) / 2
+                            ,
+                            demand_uncertainty / 2
+                        ),
+                        2 * demand_uncertainty)
+                    )^2
+                    +(MIN(
+                        MAX(
+                            LN(
+                                IF(
+                                    upper_bound_demand_acc<1.2*default_median,
+                                    upper_bound_demand_acc,
+                                    1.2*default_median
+                                )/
+                                IF(
+                                    lower_bound_demand_acc<1.2*default_median,
+                                    lower_bound_demand_acc,
+                                    1.2*default_median
+                                )
+                            )/2,
+                            beta_c/2
+                        ),
+                        2*beta_c
+                    )
+                    )^2
+                    +
+                    beta_t^2)
+                )
+            )
+    '''
+
+    uncertainty_lookup = {
+        'best': {
+            'baseline': {
+                'beta_t': .2,
+                'beta_c': .1
+            },
+            'non-baseline': {
+                'beta_t': .4,
+                'beta_c': .2
+            }
+        },
+        'very_good': {
+            'baseline': {
+                'beta_t': .25,
+                'beta_c': .15
+            },
+            'non-baseline': {
+                'beta_t': .45,
+                'beta_c': .25
+            }
+        },
+        'good': {
+            'baseline': {
+                'beta_t': .3,
+                'beta_c': .2
+            },
+            'non-baseline': {
+                'beta_t': .5,
+                'beta_c': .3
+            }
+        },
+        'poor': {
+            'baseline': {
+                'beta_t': .4,
+                'beta_c': .25
+            },
+            'non-baseline': {
+                'beta_t': .55,
+                'beta_c': .35
+            }
+        },
+        'very_poor': {
+            'baseline': {
+                'beta_t': .5,
+                'beta_c': .3
+            },
+            'non-baseline': {
+                'beta_t': .6,
+                'beta_c': .4
+            }
+        }
+    }
+
+    performance_rating = (
+            performance_rating if performance_rating == 'baseline'
+            else 'non-baseline'
+    )
+
+    beta_c = uncertainty_lookup[quality_rating][performance_rating]['beta_c']
+    beta_t = uncertainty_lookup[quality_rating][performance_rating]['beta_t']
+
+    return min(1.1 * default_beta, max(.9 * default_beta, math.sqrt(min(max(math.log(min(upper_bound_demand_disp, 1.2 * default_median) / min(lower_bound_demand_disp, 1.2 * default_median))/2, demand_uncertainty / 2), demand_uncertainty * 2)**2) + min(max(math.log(min(upper_bound_demand_acc, 1.2 * default_median) / min(lower_bound_demand_acc, 1.2 * default_median))/2, beta_c / 2), 2 * beta_c)**2 + beta_t**2))
